// FrontEnd/src/app/models/Reports.model.ts
export class Report {
    roomNum: number;
    name: string;
    phoneNum: number;
    theProblems: string;
    Requre: string;
    title: string;
}

export interface report {
    roomNum: number;
    name: string;
    phoneNum: string;
    theProblems: string;
    Requre: string;
    title: string;
}